ordering_agent/agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
fastapi==0.116.1
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.2.3
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-pydantic==0.6
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-core==0.3.71
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.8
mcp==1.12.1
mcp-use==1.3.9
multidict==6.6.3
oauthlib==3.3.1
openai==1.97.1
orjson==3.11.0
packaging==25.0
posthog==6.3.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
scarf-sdk==0.1.2
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
sse-starlette==2.4.1
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
